re-organize project, add new list feature
still not happy with the docker compose UX, need to finalize .env stuff, etc.

The list feature is not yet implemented

also need to add dockerfile for frontend

also need to make sure `mysql-data` dir is completely within the docker volume so i never accidentally delete it

also, i'd hate it if there are incompatibilities with previous versions.